rename octree_stream to octree.py. bye bye old implementation :'[
